:building_construction:&&:sparkles: Refractor the codes structure and add the redefine check
